Remove InstalledVersion
It seems like InstalledVersion, while being fancy, is not really
useful. In detail, it is quite often "wrong" / we cannot easily
identify the installed version by parsing the output of go get.
There would be the possibility to make this work, but it does not
seem to provide a lot of benefit to actually do that work.