truncate config file before writing new config
This commit fixes #6. If packages are removed, the config file gets
shorter. So far, we simply overwrote the config / state file, which
resulted in garbage at the end of the new content, which was related
to the old content.
To fix this, we now truncate (clear) the file before writing the
new contents to the file.