Refactor cmd structure and add brew cask handling
This commit introduces the handling of brew casks, which can be
installed by providing "cask" as the first package name.

It does so by refactoring the command structure and handlers. Now, the
handlers need to have a `Command() *cobra.Command` method, where all the
subcommands (list, upgrade, install, remove) will be attached to.
This allows the different handlers to have different flags, which is
used for the cask handling (`--cask`). It also switches how
packa is called, instead of `packa install go [pkg]`, `packa go install
[pkg]` is now used.

Signed-off-by: Ramon Rüttimann <ramon.ruettimann@gmail.com>